refactor: move ObjectMapperException.class to exception package
