+++ nextstep-mvc/src/main/java/nextstep/mvc/tobe/exception/ObjectMapperException.java
@@ -1,4 +1,4 @@
-package nextstep.mvc.tobe;
+package nextstep.mvc.tobe.exception;
 
 public class ObjectMapperException extends RuntimeException {
     public ObjectMapperException() {
